fix(kb,taskrace): Cleanup+fix for Invalid argument index in format
Cleanup of AddCompleteTransfersv2 and added psrc := nil; like I have had
in my old incfiller code.

On taskrace is an error: Invalid argument index in format "EMPTY PS1 %s
: LastChange(%d) > newdir_max_empty(%d)
It was already before there because LastChanged is TDateTime, so can't
just be output with %d. Not sure if we should use the output with
SecondsBetween or just the value from Lastchanged with
DateTimeToStr(d.LastChanged)