fix(setdir): Some fixes for last !setdir changes
Fixed the error message when section is not found (it was saying site not found)

Added a hint for the user that the section must exists in slftp.precatcher

Also disabled the check if the dir is not specified, so people are able to remove section if they removed it from slftp.precatcher.

madjik bozo out!