fix(!tweak): Function use lowercase for input values now
To fix #78 we have to lowercase the input values.
I also changed the values 'IRCNick' to 'ircnick' in sourcefiles.
Only the site property `property IRCNick` is no lowercase now.
If someone reports a problem which occur by using lowercase, we can fix
it later.
But every value for my sites.dat file is lowercase...

This should close #78.