feature(IrcRuleAdd): Output ID from added rule, Use same msg for rule cmd
Someone asked if we could output the ID from a new added rule. You don't
need to do !rules to find out if you forget something or made a
mistake in your rule.
I changed the text output in every function for rules to the same, so
the err msg are the same now.
We don't need queue_lock.Enter and queue_lock.Leave for copying rules.

Closes #91.