refactor(parsetvmazeinfos): Change try..Except to where it's needed, Check if Network->Country exist
I removed the 'genre is null' error msg, because it's not needed
anymore. I changed the try..exception block to thetvdb genre parsing
because xml.LoadFromStream() still creates 'In 'stream:' (line 1 pos
55): Expected whitespace' error msg. (See #28)
Check for Network->Country is added, so we only parse it if it
exist, else we return with unknown. (Mentioned in #70)