refactor(!country): Only acronyms can be set for country
You can only use acronyms to set a country for a site. It'll show
Countryname for countrycode on !site YOURSITE.
Countryname is also shown while adding country to a site.
It's needed for new !routeset cmd and you can't fuckup your countrys, so
!findcountry will work better and you have to use same country for same
country. No more .nl, .netherland and netherland possible for country.