fix(IrcNoLoginMSG): Support for GLFTPD only + skip admin_site
Bypassing login msg only works on glftpd sites - you insert a '-' as
first char on your password.
If you do this on drftpd site it says you have a wrong password.
So it makes only sense to set NoLoginMSG on glftpd sites. With old code
also admin_site was shown.