change(precatcher): Changed triggers to do a MP3 genre lookup on the site bot annouce line
Now instead of checking if section name contains MP3, we check if the
section is part of TMP3Release.

This enable genre lookup for other music sections like FLAC.

Also we now skip the attepmt to get the genre from NEWDIR event.
Sites doesn't know the genre of a mp3 release until the first file is
sent.

Some sites display a line after the first file is sent, maybe we should
handle an catcher event (UPDATE?) and allow it to be parsed for genre ?