refactor(irc, news): Real CTCP Support, admin_net remove, Month as number in msgsystem
Old code for CTCP was broken and didn't worked, so I repaired it and
added some more commands. You even can ping your slftp and see if it's
still alive.
Added some comments to understand how incoming msgs looks like and how
it's processed.
I even removed the admin_net stuff as it's deprecated now. We have a
message system which is much better!
Fixed also the bug madjik had with locale month names in message system.
It just shows the number of the month instead the name.

For CTCP see:
http://www.irchelp.org/protocol/ctcpspec.html
https://tools.ietf.org/html/draft-oakley-irc-ctcp-01