refactor(taskrace,stats): Reformatting and fix code from Spinat
Reformatted the code as it was really bad formatted!
We even don't need to announce stats to irc if we didn't raced the file,
makes no sense to show info if we got kicked because of slowkick or
something like this.
No need to convert int64 to double for stats, converting it to string
and then convert it to integer again.

NOTE: was a mistake to add upcoming SlftpNewsAdd() command in here...