feat(startup): List missing files + 0byte file check
If slftp misses any important files which is not autogenerated, it will
throw a message at startup which shows which file is missing (or 0
byte).
For autogenerated files, it only checks for filesize and throws a
message when a 0 byte file was detected.
Whenever something is not correct, it won't start.

NOTE: not tested with a real 0 byte file - it might won't work as it is
implemented now as those files maybe are bigger than 0 byte

This closes #189