feat(Indy10): Add latest Indy to libs
Added latest Indy10 version to slftp libs folder.
I chosed dirname with revision, so that we always know which version we
use.
Would be better if we could use * in makefiles, so that we don't need to
change it when we update Indy folder.
Created a cleanup.sh to delete (hopefully) unneeded Indy files. I also
hope that all relevant files are there as Indy comes with several
folders etc. :D

Indy mirror: https://indy.fulgan.com/