replace(TSafeInteger): Use Indy's TIdThreadSafeInt32
replaces TSafeInteger class with Indy's TIdThreadSafeInt32 class
added TIdThreadSafeInt32WithEvent which can handle events which is
only needed for queuenumber

Old TSafeInteger class didn't allow values < 0 which is a bad idea in
my eyes because you don't see if anything is wrong in code because it
prevents showing negative (=bad) numbers. Normally it should never
occur that a value is negative because it's first increased in
TPazoTask.Create and in TPazoTask.Destroy decreased.