fix(addtvmaze_channel): Prevent unnecessary lookups
As first it searches in database if a show with ID from !addtvmaze
channel already exists, it will only create a task if it does not exist
or if the last update is older than the value given in
days_between_last_update
Also improved the stripping of scene tags with the use of existing
getShowValues procedure.

Also a mem leak in saveTVInfos was fixed, getTVInfoByShowID creats a
TTVInfoDB object and returns it to the caller but the caller didn't
freed it! But the call to getTVInfoByShowID is not needed anymore.