fix(queueunit): free slots faster
Transfer A->B assigns a RaceTask to A and a WaitTask to B
Previous queue behaviour would first check if the WaitTask on B was
finished and then check if the RaceTask on A was finished.
If the RaceTask on A is finished it will notify the WaitTask on B that
it is also finished.
This led to a delay in freeing of slots, because the slot on A would be
released, but the slot on B would only be released on the next
iteration of the queue. In the meantime no new RaceTasks between A and
B would be started because while slot on A was free, slot on B was not
free yet.
As such switch the checks around so RaceTask is checked for completion
first and sets WaitTask to done which is checked next and frees up the
slots on both sites.