refactor(rules): Rename Hitelesit to Verify
Make it more clear what it is that hitelesit actually does and stands
for.
It's called when new rules are added to verify whether the supplied
values are valid values for a given rule.
Example: if mp3language = FOOBAR then DROP
  will fail because FOOBAR is not a known mp3 language tag in our setup

If no Verify function is provided for a rule it accepts all inputs.