fix(pazo): Improve MKDIR for subdirs
Previously we could run into problems if the dirlist for a subdir was
executed before the maindir was MKDIRd. Since the maindir wasn't created
yet the MKDIR+RACE tasks for the subdir wouldn't be issued.
Since QueueSort grants higher priority to mkdir tasks than race tasks
and to mkdir tasks for maindirs vs mkdir tasks for subdirs we can
just remove this check and it'll still work as expected.