fix(dirlist): tasks being stuck - deadlock with new locking in TPazoSite.ParseDirlist
when pazo-level critical section was removed there was race condition introduced
because the lock remaining was on dirlist-level. In function Tuzelj a FindDirList is invoked
which would lock another TDirList after which another thread could enter TPazoSite.ParseDirlist
and try to enter the lock for that dirlist -> deadlock.

Solved by invoking Tuzelj outside of the dirlist_lock.
Also I removed the new FParseDupe_cs and replaced it by a dirlist_lock, because
the parsedupe code actually requires a dirlist_lock.