perf(queue): relogin slots via IdleTask, add exception logging, refactor queue iteration
As we want all the slots of a site to be ready for action, IdleTasks are also created
for slots that are not ssOnline if the site is in WorkingStatus sstUp.
Also add exception logging for exceptions that have been hidden before.
Remove the extra step of marking tasks to be deleted first and then deleting them in
a second step - just delete the tasks in 1 step.
Also do only 1 lock over both remaining
steps in queue iteration to avoid AddTask adding tasks which are ready right away
between the steps which we would then need to iterate as well. also that distorts
the number of tasks shown in the queue summary at the bottom of the console window.