fix(taskrace): loop sending the same already existing file to a site
Fix endless retries to send already existing files to a site. That could happen when all slots
of a site are taken by race tasks, so none are free for dirlist tasks. We would try to send a
file that is already there and then rely on x-dupe to show this and we would then parse x-dupe
and create a dirlistentry for that file. That would prevent us from sending the file again
even though there has been no dirlist. But there are files which are not part of the x-dupe
output. In glftpd default config, files like sfv, mp4 are not to be included in x-dupe. Then we
won't know the file is already there and create race tasks over and over again.
That is fixed by invoking ParseDupe for the file we're trying to send, in case a dupe message
occurs and not solely rely on x-dupe.

Some other issues are also addressed, see below.

TaskRace:
Invoke ParseDupe whenever we get a dupe message (and determine whether the file is complete
or incomplete on the destination site).

Add handling of PRET responses. We can get errors like dupe messages or 'File unavailable' and so
on and not waste time on trying to send STOR or RETR.

In general we ignored quite some error responses. There are these case lastResponseCode of - blocks
where we handle specific response codes. But if we don't know the message, we ignore the error and
continue execution. That inevitably leads to errors for subsequent commands and wastes time.
Change the case .. of parts so that the "unhandled response" message doesn't only come up for
response codes that we do not handle, but also for unhandled messages of handled response codes.
Now these error are no longer ignored.
Also add some handling of so far unhandled response messages.

Added a few irc error messages and debug messages for errors that would else remain undetected by
the user. E.g. parts like "421 Timeout (10 seconds): closing control connection." There is a goto
TryAgain, but the user may wants to know which site is having issues and blocks slots.

Remove ParseDupe call from places where we don't know that the file is even there. For example
'SFVFile still transferring'. That is no dupe message.

Add ParseDupe where we know, the file is there: When we successfully issued STOR and RETR, we
know, the file is being uploaded, because we do it.

Do ps2.SetFileError when we know, the file cannot be transfered:
- 'does not exist in the sfv'
- 530 Access denied
- otherwise unhandled 553 responses which means "Requested action not taken. File name not allowed"
- 'CRC-Check: Not in sfv!'

Remove useless ps2.ParseDirlist after transfer. That response does not contain a dirlist.

pazo:
Create race tasks also from ParseDupe, not only from ParseDirList.
There is no need to wait for a dirlist to create a RaceTask.
Also consider "skip being uploaded files" setting.

FIX: Invoke RegenerateSkipList for new DirListEntries in ParseDupe. This solves the issue
of files to be skipped being traded anyway. E.g. FTPRush-screwed files where sometimes being
traded.

Remove unnecessary RemovePazoRace calls in TPazoSite.ParseDupe and TPazoSite.ProcessXDupeResponse.