fix(taskrace): remove auto-set of legacydirlist in case of 'PRET required' FTP response
This was only done at one place. I decided to remove it, to provide consistent behaviour.
It was dependent on whether the first transfer of the site was an upload or a download. In
only one of these cases, legacydirlist would be set automatically.
Also there are drftpd sites that don't require legacydirlist (newer versions).