fix(dirlist): don't apply filename checks to dirs
The IsValidFilename check enforces a minimum length of 5 chars for its
arguments under the assumption that every file will at least have
<something>.<extension> with something being at least 1 char, the dot,
and an extension of at least 3 chars.
This is also applied to directories in the dirlist which leads to
skipping dirs like 2020, 0101, S01, etc. which are frequently
encountered when using the indexing feature leading to lots of dirs
being skipped and not indexed.