fix(dirlist): fix no such directory error flood for subdir dirlists
If a subdir is not available it could happen that we run into a loop
of dirlist errors (like no such directory). That is now handeled by
setting d.need_mkdir := true. Also if it happens for the main dir,
MarkSiteAsFailed is invoked, same as when this happens for STOR command.

Also if MKDIR is slow, it can happen that the dirlist happens too early
so that the dir is not there yet. Don't treat this as error, but just log
a message to the log file.