fix(taskrace): avoid 'tell your developer' messages
since there are many scripts that produce messages that are unknown to us and
also there are untested ftpds, there are many unnecessary 'tell your developer'
messages spamming the admin channel. Replace these by simple FXP-errors which
are sent to the error channel.
ERROR No Such File handling is moved to a local procedure to be used for both RETR and PRET RETR