feat(siteSoftware): Better detection of site software, detect software version & Pure-FTPd support
Use site messages from login as well as STAT and SITE VERS cmd to determine the site software in a
more secure way than guessing by FEAT list which was used before and still is used as a fallback.
Also add Pure-FTPd site software which requires 'PBSZ 0' to be sent for secure connections.