feat(dirlist): separate global skip for directories
instead of haveing only one global_skip setting in slftpi.ini for both files and directories,
create a new setting global_skip_dir which is only applied to directories. existing setting
global_skip is now only applied to files.

It's no longer needed to consider autoindex-dirlists because that was only needed because
of the min-length check in IsValidFilename.