fix(autodirlist): fix strange race condition when readding autodirlist tasks
When a site is being !setdown and !bnctest it will readd autodirlist tasks. this did
not (always) work because of some race condition which i don't completely understand.
However adding a queue lock to when there is checked if there's already a autodirlist
task in the queue seems to solve this issue. And this lock should be there anyway.
Also added this to same patterns for autoindex, autonuke, autobnctest, autorules.