fix(reqfiller): fix reqfill cmd on incomplete releases and show proper info on error
Only do the comparison for number of files when the pazo is ready. This to avoid sending
'site reqfilled' before the release is complete when files are still being transfered.
dirlist.done only get the file count and does not check if the file is actually complete
or still being transfered.

Also show a proper info with error if the pazo becomes ready without the rls getting
completed.