fix(queue): fix assign of slots that are not online
Refactor finding of a free slot and thereby fix a bug which caused not online slots
to be assigned for a task. This happened because in the loop the variable is always
assigned and never set to nil, then the last slot will be used to assign the task, even
if it is not online.