fix(autologin): fix not working autologin feature
When creating a task and the corresponding site is not yet in the sites list,
the task will be set to ready error right away and never be executed.
Because the login task for autologin is being created while still in TSite.Create,
the site is not yet in the sites list at this moment, therefore the autologin feature
did not work.
Change this to add a login task at SitesStart. Also remove the no longer needed
procedure AddLoginTask.