fix(openssl): fix app crashes when calling SSL shutdown and free in parallel
I've seen app crashes when a site has issues and TslTCPSocket.DisconnectSSL would be invoked
frequently. It would be "double free or corruption" errors, also TurnToSSL popped up in the
stack trace. It seems this can happen if SSL_shutdown and SSL_free are invoked at the same time
(on different SSL objects). Therefore put those in the sltcp_lock.