fix(race): fix possible deadlock with ReleaseSlotsAssignmentLock
When trying to RemoveActiveTransfer if the site is overloaded this could take some time and
if the queue clean for example would do something with this task in the meantime, this
might not work and cause an exception instead:
TSite(TPazoRaceTask(todotask).ssite2).ReleaseSlotsAssignmentLock;
This would lead to the lock never being released. Fix this by using the same TSite variable that
is already being used when acquiring the lock.